// apps/backend/src/services/AppService.d.ts
import { BaseService } from './BaseService.js';
import { InfrastructureService } from './InfrastructureService.js';
import { ClusterService } from './ClusterService.js';
import { BuilderService } from './BuilderService.js';
import type { LocalDB } from '../lib/db.js';
import type { DeploymentMetadata } from '../lib/types.js';
import { Server as SocketServer } from 'socket.io';
export declare class AppService extends BaseService {
    private infra;
    private clusters;
    private builder;
    constructor(db: LocalDB, infra: InfrastructureService, clusters: ClusterService, builder: BuilderService);
    getAll(io?: SocketServer): Promise<DeploymentMetadata[]>;
    private sanitize;
    deploy(config: any, io?: SocketServer): Promise<DeploymentMetadata>;
    updateModules(id: string, modules: string[], io?: SocketServer): Promise<{
        modules: string[];
        status: "deploying";
        id: string;
        name: string;
        clusterId: string;
        strategy: "helm" | "native";
        appType?: "odoo" | "wordpress" | "nextcloud" | "audiobookshelf" | "prometheus" | "traefik";
        webRepo?: string;
        webTag?: string;
        dbRepo?: string;
        dbTag?: string;
        url?: string;
        isExposed?: boolean;
        exposureUrl?: string;
        lastLogPath?: string;
        storage?: Record<string, string>;
        vpnEnabled?: boolean;
        vpnProtocol?: "wireguard" | "openvpn";
        vpnConfig?: string;
        vpnDedicatedIp?: string;
        temporalWorkflowId?: string;
    }>;
    listPods(id: string): Promise<{
        pods: any;
        namespace: string;
    }>;
    getHelmStatus(id: string): Promise<string>;
    getDiagnostics(id: string): Promise<string>;
    delete(id: string, io?: SocketServer): Promise<void>;
    resizeDisk(id: string, storage: any, io?: SocketServer): Promise<{
        storage: any;
        status: "deploying";
        id: string;
        name: string;
        clusterId: string;
        strategy: "helm" | "native";
        appType?: "odoo" | "wordpress" | "nextcloud" | "audiobookshelf" | "prometheus" | "traefik";
        webRepo?: string;
        webTag?: string;
        dbRepo?: string;
        dbTag?: string;
        url?: string;
        isExposed?: boolean;
        exposureUrl?: string;
        lastLogPath?: string;
        modules?: string[];
        vpnEnabled?: boolean;
        vpnProtocol?: "wireguard" | "openvpn";
        vpnConfig?: string;
        vpnDedicatedIp?: string;
        temporalWorkflowId?: string;
    }>;
}
//# sourceMappingURL=AppService.d.ts.map